replace tests and add notification
