+++ github-clone/src/api/octokit/index.ts
@@ -1,5 +1,5 @@
 import { Octokit } from "octokit";
 
 export const octokit = new Octokit({
-  auth: "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",
+  auth: "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",
 });
