+++ github-clone/src/components/Navbar/index.tsx
@@ -13,7 +13,11 @@ import {
 } from "../../store/filters";
 import { getUsersEvent } from "../../store/users";
 
-export const Navbar: React.FC = () => {
+interface NavbarProps {
+  isLoading: boolean;
+}
+
+export const Navbar: React.FC<NavbarProps> = ({ isLoading }) => {
   const [showNavbar, setShowNavbar] = useState<boolean>(true);
   const toggleNavbar = () => setShowNavbar(!showNavbar);
 
@@ -45,6 +49,7 @@ export const Navbar: React.FC = () => {
         </div>
         <Form.Item label="Name">
           <Input
+            disabled={isLoading}
             placeholder="Name"
             aria-label="name-input"
             value={userName}
@@ -53,6 +58,7 @@ export const Navbar: React.FC = () => {
         </Form.Item>
         <Form.Item label="Followers count from">
           <Input
+            disabled={isLoading}
             type="number"
             aria-label="followers-input"
             placeholder="Followers count from"
@@ -62,6 +68,7 @@ export const Navbar: React.FC = () => {
         </Form.Item>
         <Form.Item label="Created by">
           <Input
+            disabled={isLoading}
             placeholder="Created by"
             aria-label="date-input"
             type="date"
@@ -70,7 +77,9 @@ export const Navbar: React.FC = () => {
           />
         </Form.Item>
 
-        <Button onClick={getUsersEvent}>Найти</Button>
+        <Button onClick={getUsersEvent} disabled={isLoading}>
+          Найти
+        </Button>
       </div>
     </div>
   );
