+++ github-clone/src/pages/Main/style.module.scss
@@ -1,5 +1,6 @@
 .main__content {
   width: 100%;
+  overflow: auto;
 }
 
 .main__table-container {
